Add docs-components library
Adds a new library of components to be used in documentation. These
consume the data produced by part of Storybooks webpack config to allow
us to display the information in a format people are more familiar with
from the GovUK Design System.

In the long run, these components should detect the presence of
Storybook and wrap Storybook's addon-docs components which they
effectively replace. This would allow the MDX files to be displayed in
Storybook as well as a documentation website.