docs-components: Use PrismJS for all highlighting
Switches Highlight.js for PrismJS when applying syntax highlighting to
HTML. This simplifies the code bringing it into alignment with React. It
also means that visually there is more symmetry between the two.

It does come at the cost of reduced contrast in the colours. - We may
want to re-create the old colour scheme with PrismJS.