BREAKING: docs-components: Update for Storybook v9
Also updates prettier to v3, this prevents rendering code snippets on
the server due to prettier removing its synchronous API. There might be
a workaround in React v19 though.

BREAKING: Must only be used with stories defined in CSF v3.

REEGRESSION: Code snippets are not populated on the server but only
after the client has hydrated.